Update scripts to import into cgds_hg1938 database (#1046)
Update scripts and import-tool to import into cgds_hg1938 database

+++ src/import-tool.cc
@@ -109,6 +109,11 @@ static string KILL_DEVDB_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devd
 static string DEVDB_IMPORT_IN_PROGRESS_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-in-progress");
 static string DEVDB_IMPORT_KILLING_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-killing");
 static string DEVDB_IMPORT_LOG_FILENAME(IMPORT_LOG_BASEDIR + "/devdb-importer.log");
+static string START_HGNC_1938_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-1938-import-start-request");
+static string KILL_HGNC_1938_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-1938-import-kill-request");
+static string HGNC_1938_IMPORT_IN_PROGRESS_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-1938-import-in-progress");
+static string HGNC_1938_IMPORT_KILLING_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-1938-import-killing");
+static string HGNC_1938_IMPORT_LOG_FILENAME(IMPORT_LOG_BASEDIR + "/hgnc-1938-importer.log");
 
 /* used for command line argument checking */
 static vector<string> RECOGNIZED_IMPORTERS;
@@ -222,6 +227,7 @@ void initialize_static_objects() {
     RECOGNIZED_IMPORTERS.push_back("triage");
     RECOGNIZED_IMPORTERS.push_back("hgnc");
     RECOGNIZED_IMPORTERS.push_back("devdb");
+    RECOGNIZED_IMPORTERS.push_back("hgnc1938");
     initialize_static_managed_importers();
     RECOGNIZED_COMMANDS.push_back("start");
     RECOGNIZED_COMMANDS.push_back("kill");
@@ -246,7 +252,7 @@ string get_hostname() {
 void print_usage(string program_name) {
     vector<char> program_name_vector = character_vector_from_string(program_name);
     cerr << "Usage: " << basename(&program_name_vector[0]) << " importer_name command [extra_arguments]" << endl;
-    cerr << "       importer_name must be \"genie\" or \"triage\" or \"hgnc\" or \"devdb\"" << endl;
+    cerr << "       importer_name must be \"genie\" or \"triage\" or \"hgnc\" or \"devdb\" or \"hgnc1938\"" << endl;
     cerr << "       valid commands:" << endl;
     cerr << "           start : requests that an import run begins as soon as possible - this may wait for an import in progress to finish before starting" << endl;
     cerr << "           kill : requests that any import in progress be halted and that any requested start be canceled" << endl;
@@ -446,6 +452,9 @@ int request_importer_start(string importer) {
     if (importer == "devdb") {
         return request_importer_start(importer, START_DEVDB_IMPORT_TRIGGER_FILENAME, KILL_DEVDB_IMPORT_TRIGGER_FILENAME, DEVDB_IMPORT_KILLING_FILENAME);
     }
+    if (importer == "hgnc1938") {
+        return request_importer_start(importer, START_HGNC_1938_IMPORT_TRIGGER_FILENAME, KILL_HGNC_1938_IMPORT_TRIGGER_FILENAME, HGNC_1938_IMPORT_KILLING_FILENAME);
+    }
     cerr << "Error : unrecognized importer " << importer << " during attempt to start import." << endl;
     return 1;
 }
@@ -462,6 +471,9 @@ int request_importer_kill(string importer) {
     }
     if (importer == "devdb") {
         return create_trigger_file(KILL_DEVDB_IMPORT_TRIGGER_FILENAME);
+    }
+    if (importer == "hgnc1938") {
+        return create_trigger_file(KILL_HGNC_1938_IMPORT_TRIGGER_FILENAME);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to kill import." << endl;
     return 1;
@@ -514,6 +526,9 @@ int report_importer_status(string importer) {
     }
     if (importer == "devdb") {
         return report_importer_status(importer, START_DEVDB_IMPORT_TRIGGER_FILENAME, KILL_DEVDB_IMPORT_TRIGGER_FILENAME, DEVDB_IMPORT_IN_PROGRESS_FILENAME, DEVDB_IMPORT_KILLING_FILENAME, DEVDB_IMPORT_LOG_FILENAME);
+    }
+    if (importer == "hgnc1938") {
+        return report_importer_status(importer, START_HGNC_1938_IMPORT_TRIGGER_FILENAME, KILL_HGNC_1938_IMPORT_TRIGGER_FILENAME, HGNC_1938_IMPORT_IN_PROGRESS_FILENAME, HGNC_1938_IMPORT_KILLING_FILENAME, HGNC_1938_IMPORT_LOG_FILENAME);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to report status." << endl;
     return 1;
@@ -583,6 +598,9 @@ int report_importer_log(string importer, vector<string> & extra_args) {
     }
     if (importer == "devdb") {
         return report_importer_log(importer, DEVDB_IMPORT_LOG_FILENAME, extra_args);
+    }
+    if (importer == "hgnc1938") {
+        return report_importer_log(importer, HGNC_1938_IMPORT_LOG_FILENAME, extra_args);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to report log." << endl;
     return 1;
